flaskAIStargpt-master/pyproject.toml
[tool.poetry]
name = "flaskaistargpt"
version = "0.1.0"
description = "ChatGPT Clone using Flask"
authors = ["Kamilou Issaka El Kémal"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "2.3.3"
openai = "0.27.10"
python-dotenv = "1.0.0"
chatterbot-corpus = "^1.2.0"
chatbot = "^1.5.2b0"
pyyaml = "3.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
